Add initial SQL prepared statement calls
Added the initial SQL prepared statement calls to the DefaultDatabase
implementation. They need to be tested to see if they work, but adding
credentials to VCS would be insecure.

In addition, one method has not been implemented.

+++ src/main/kotlin/latesco/core/data/Asset.kt
@@ -36,4 +36,4 @@ package latesco.core.data
  * @param name      The human readable name for an asset.
  * @param symbol    The symbol which represents the asset.
  */
-data class Asset(val uid: Long, val name: String, val symbol: String)
+data class Asset(val uid: Int, val name: String, val symbol: String)
